requirements.txt
crewai==0.114.0

streamlit>=1.35.0,<1.45.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.7.0,<3.0.0

requests>=2.30.0,<3.0.0
httpx>=0.27.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
tenacity>=8.2.0,<9.0.0
rich>=13.0.0,<14.0.0

tavily-python>=0.5.0,<0.6.0
groq>=0.9.0,<1.0.0